requirements-dev.txt
build==0.10.0
twine==5.1.1
bump2version==1.0.1
pytest==8.1.1
pytest-cov==4.1.0
ruff==0.6.9
cloudevents==1.11.0
mkgendocs==0.9.2
mkdocs-material==9.5.40
mkdocs==1.6.1
PyYAML==6.0.1